Adicionando dialog para confirm de ações sem volta

+++ viajatcc-frontend/src/app/components/bus/fornecedores/fornecedores.component.ts
@@ -1,6 +1,8 @@
 import { Component } from '@angular/core';
 import { BusService, supplier } from '../../../services/bus.service';
 import { Router } from '@angular/router';
+import { ConfirmationDialogComponent } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-fornecedores',
@@ -12,30 +14,33 @@ export class FornecedoresComponent {
   suppliers: supplier[] = []
   isLoading = true;
   error = '';
-  constructor(private busService: BusService, private router: Router ) {
+  constructor(private busService: BusService, private router: Router, private dialog: MatDialog) {
     this.busService.getSuppliers().subscribe({
       next: (response) => {
         this.suppliers = response
         this.isLoading = false;
-    },
+      },
       error: (err) => {
         this.error = 'Erro ao carregar os fornecedores';
         this.isLoading = false;
         console.error(err);
       }
-  })
+    })
   }
   viewSupplier(id: string) {
     console.log(id);
-    
+
   }
-  deleteSupplier(id: string) {
-    this.busService.deleteSupplier(id).subscribe({
-      next: (response)=> {
-        this.router.navigate(['/home/bus/list'])
-        console.log('excluido');
-        
-      } 
-    })
+  async deleteSupplier(id: string) {
+    const confirm = await this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja excluir este fornecedor?'
+    }).afterClosed().toPromise()
+    if (confirm) {
+      this.busService.deleteSupplier(id).subscribe({
+        next: (response) => {
+          this.router.navigate(['/home/bus/list'])
+        }
+      })
+    }
   }
 }
